taito/flstory.cpp: Renamed Fairyland Story sets and cleaned up a little.
* Renamed the clone set - it's an earlier version, not a version
  specifically for Japan (addresses MT08579).
* Consolidated source files, replaced literal tags, and removed some
  vestigial code from the old MCU simulation.
