+++ src/mame/mame.lst
@@ -42161,7 +42161,7 @@ fgoala                          // MF  (c) 1979 Taito Corporation
 
 @source:taito/flstory.cpp
 flstory                         // A45 (c) 1985 Taito Corporation
-flstoryj                        // A45 (c) 1985 Taito Corporation (Japan)
+flstoryo                        // A45 (c) 1985 Taito Corporation (Japan)
 onna34ro                        // A52 (c) 1985 Taito Corporation (Japan)
 onna34roa                       // A52 (c) 1985 Taito Corporation (Japan)
 rumba                           // A23 (c) 1985 Taito Corporation (Japan)
